pyproject.toml
[project]
name = "surgeontrainer-api"
version = "2.0.0"
description = "FastAPI backend for SurgeonTrainer - Comprehensive patient and surgical case management"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.13.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = []
